+++ src/OpenSimCreator/Graphics/OpenSimDecorationOptions.cpp
@@ -154,6 +154,16 @@ void osc::OpenSimDecorationOptions::setShouldShowContactForces(bool v)
     SetOption(m_Flags, OpenSimDecorationOptionFlags::ShouldShowContactForces, v);
 }
 
+bool osc::OpenSimDecorationOptions::getShouldShowForceLinearComponent() const
+{
+    return m_Flags & OpenSimDecorationOptionFlags::ShouldShowForceLinearComponent;
+}
+
+void osc::OpenSimDecorationOptions::setShouldShowForceLinearComponent(bool v)
+{
+    SetOption(m_Flags, OpenSimDecorationOptionFlags::ShouldShowForceLinearComponent, v);
+}
+
 void osc::OpenSimDecorationOptions::forEachOptionAsAppSettingValue(const std::function<void(std::string_view, const Variant&)>& callback) const
 {
     callback("muscle_decoration_style", GetMuscleDecorationStyleMetadata(m_MuscleDecorationStyle).id);