+++ CHANGELOG.md
@@ -5,8 +5,12 @@ on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Upcoming Release]
 
+- The model editor's/simulator's 3D visualizers now have an experimental `Show > Forces' Linear Component`
+  visualization option, which adds arrows indicating how each `OpenSim::Force` adds its
+  linear force component to each body in the model. This can be useful for debugging model
+  creation or `ExternalForce`s
 - The model editor UI now has experimental support for viewing `OpenSim::Function` curves. This
-  is currently exposed as an eye icon in the property editor panel
+  is currently exposed as an eye icon in the property editor panel (#695)
 - Selecting an `OpenSim::Joint` that has `OpenSim::PhysicalOffsetFrame`s for both its parent
   and child frames now shows a 3D manipulation gizmo that lets you move the joint center
   without moving anything else in the model (#159)